@@ -1,16 +1,19 @@
 plugins {
     id 'java'
 }
-
-group 'com.thoughtworks'
+group 'com.thoughtworks.bootcamp'
 version '1.0-SNAPSHOT'
-
-sourceCompatibility = 1.8
-
+sourceCompatibility = JavaVersion.VERSION_11
 repositories {
     mavenCentral()
 }
-
+test {
+    useJUnitPlatform()
+}
 dependencies {
-    testCompile group: 'junit', name: 'junit', version: '4.12'
+    compileOnly 'org.projectlombok:lombok:1.18.6'
+    annotationProcessor 'org.projectlombok:lombok:1.18.6'
+    testCompile 'org.mockito:mockito-core:1.+'
+    testCompile 'org.junit.jupiter:junit-jupiter:5.4.1'
+    testCompile 'org.assertj:assertj-core:1.0.0'
 }
